requirements.txt
biopython==1.83
fair-esm==2.0.0
networkx==3.0
numpy==1.24.1
optuna==3.6.1
pandas==2.0.3
pytorch-lightning==2.2.3
rdkit==2023.9.5
scikit-learn==1.3.2
scipy==1.10.1
torch==2.2.2+cu118
torch_geometric==2.5.2
torchmetrics==1.3.2